+++ include/npcomp/Dialect/RefBackend/IR/CMakeLists.txt
@@ -1,0 +1,1 @@
+add_mlir_dialect(RefBackendOps refback)
