+++ include/npcomp/Dialect/RefBackend/IR/RefBackendDialect.h
@@ -1,0 +1,16 @@
+//===------------------------------------------------------------*- C++ -*-===//
+//
+// This file is licensed under the Apache License v2.0 with LLVM Exceptions.
+// See https://llvm.org/LICENSE.txt for license information.
+// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
+//
+//===----------------------------------------------------------------------===//
+
+#ifndef NPCOMP_DIALECT_REFBACKEND_IR_REFBACKENDDIALECT_H
+#define NPCOMP_DIALECT_REFBACKEND_IR_REFBACKENDDIALECT_H
+
+#include "mlir/IR/Dialect.h"
+
+#include "npcomp/Dialect/RefBackend/IR/RefBackendOpsDialect.h.inc"
+
+#endif // NPCOMP_DIALECT_REFBACKEND_IR_REFBACKENDDIALECT_H
