+++ lib/Dialect/CMakeLists.txt
@@ -2,6 +2,7 @@ add_subdirectory(ATen)
 add_subdirectory(Basicpy)
 add_subdirectory(Npcomprt)
 add_subdirectory(Numpy)
+add_subdirectory(RefBackend)
 add_subdirectory(TCF)
 add_subdirectory(TCP)
 add_subdirectory(Torch)
