+++ test/E2E/lower-to-npcomprt-abi.mlir
@@ -75,7 +75,7 @@ func @multiple_blocks(%arg0: memref<?xf32>) -> memref<?xf32> {
 
 
 // CHECK: npcomprt.global @g dense<7.000000e+00> : tensor<10xf32>
-tcp.global @g dense<7.0> : tensor<10xf32>
+refback.global @g dense<7.0> : tensor<10xf32>
 // CHECK-LABEL: func @gets_global() -> !npcomprt.tensor
 func @gets_global() -> memref<10xf32> {
 // CHECK:    %[[GMEMREF:.*]] = npcomprt.get_global @g : memref<*xf32>
@@ -83,7 +83,7 @@ func @gets_global() -> memref<10xf32> {
 // CHECK:    %[[OUTABIMEMREF:.*]] = memref_cast %[[ORIGMEMREF:.*]] : memref<10xf32> to memref<*xf32>
 // CHECK:    %[[RET:.*]] = npcomprt.from_memref %[[OUTABIMEMREF]] : memref<*xf32>
 // CHECK:    return %[[RET]] : !npcomprt.tensor
-  %0 = tcp.get_global_memref @g : memref<10xf32>
+  %0 = refback.get_global_memref @g : memref<10xf32>
   return %0 : memref<10xf32>
 }
